feat: Add `GITHUB_USE_WEBHOOK` environment variable
This new environment variable can be used to disable creating a GitHub web hook on new Insights. This is useful for local development situations where there is no public URL that GitHub can use to initialize the webhook.

The default is true so there is no change in behavior.

An error is logged if `GITHUB_USE_WEBHOOK` is true, but `PUBLIC_URL` is not set.