feat: Add Insight Links
Insight links can be attached to an Insight, and appear in the sidebar (or infobar) and link to external sites.  Links can optionally have a `name` which is displayed, or a `group` which is used to categorize and group links.

Links are editable in the Metadata tab of the Insight Editor, or they can be managed directly in the `insight.yml` file.