fix: Prevent `:iex:` suggestion from sticking around
Previously, typing `:` in the editor would trigger the auto-complete popover.  Following that with a space (e.g. if you are using a colon rather than an emoji) would remove all suggestions except the IEX logo.  This made it difficult to type colons in text as it would frequently fill in an unwanted emoji.

This change makes the `:iex:` completion work like the other emojis, so they all disappear when you follow a colon with a space.