fix: Fix anchor links to dynamic content (e.g. Markdown headings)
A new hook, `useScrollToLocation()`, automatically detects anchor links and scrolls them into view when the target element appears.  This solves the problem of opening a deeplink to a specific anchor in a Markdown document, which is rendered asynchronously.

The hook has been added to the FileViewer, so it applies to any Insights, Markdown help page, etc.