fix: Optimize Elasticsearch queries for the user profile
Previously, the `authoredInsights` and `likedInsights` resolvers would trigger Elasticsearch queries which retrieved entire Insight documents, including the contents of all Insight files.  This is obviously a huge hit to performance and network traffic, especially since the Profile page doesn't request that information in the GraphQL query.

To address this, the requested GraphQL fields are being passed down to the Elasticsearch query as the `_source` argument.  Elasticsearch now only returns the requested fields, reducing the traffic and latency.