fix: Make the Insight Editor file sidebar open initially
Ideally, the file sidebar would be closed initially on mobile, and open initially on desktop.  Unfortunately, `useBreakpointValue` defaults to `base`, so the page appears to be a mobile page on the first render.  So a conditional like this doesn't work:

```
const { isOpen: isFilesOpen, onToggle: onFilesToggle } = useDisclosure({ defaultIsOpen: !isMobile });
```

Instead of trying to workaround this, just changing it to default to open, as this is not an unreasonable default even on mobile.  We can revisit it later if necessary.