feat: Allow uploading files to subfolders via button or drag-n-drop
This change replaces the file upload drop target in the Insight Editor with the ability to upload to the root folder or any subfolder.  There is a new Upload button for each folder; it appears on subfolders when hovering over them.  Alternately, each folder (including the root) is a drop target and files can be dragged and dropped into the browser.