fix: Update `repositoryData` field when syncing Insights
The `repositoryData` field contains details needed to trigger a sync for an Insight, but previously this value was not updated, even if the Insight repository was renamed.  This continued to work thanks to GitHub's redirect functionality, but broke the previous Insight detection to avoid re-publishing unmodified files.

This change updates the repositoryData field with every successful sync.