fix: Fix editing Insight issue caused by server-side drafts
The recent changes to create new Insight drafts server-side broke the ability to edit an Insight, since there's no logic to initialize a draft from an existing Insight. This fixes the issue by allowing the client-side to create a blank draft and extend it with the contents of the Insight.