feat: Sync all collaborators to allow teams to view unlisted Insights
Currently, IEX only manages direct collaborators, e.g. users added directly to an Insight repository.  Teams and organizational users don't appear in IEX, and cannot be managed by the UI--they must be managed directly in GitHub.

However, this meant that Unlisted Insights don't appear unless you are a direct collaborator on an Insight.  This is not ideal, so this PR makes a change to the Insight sync process to include ALL collaborators (with higher than READ permissions).  These collaborators are synced to Elasticsearch and will be used to determine if an unlisted Insight can be shown.

This PR does not change the Collaborators UI within IEX--it continues to show and manage direct collaborators only.  A direct link has been added to open the Insight's repository access page in GitHub, however.